remove coordinator boiler plate
